requirements.txt
Flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
opencv-python==4.11.0.86
pillow==11.1.0
pytesseract==0.3.13
python-dotenv==1.0.1
tesseract==0.1.3
